implemented database clear --last

+++ azotea/database.py
@@ -39,13 +39,39 @@ from . import  *
 # Module global functions
 # -----------------------
 
+def latest_batch(connection):
+    '''Get Last recorded batch'''
+    cursor = connection.cursor()
+    cursor.execute('''
+        SELECT MAX(batch)
+        FROM image_t 
+        ''')
+    return cursor.fetchone()[0]
 
-def dbase_do_backup(ccomment):
+
+def dbase_do_backup(comment):
     tstamp = datetime.datetime.utcnow().strftime(".%Y%m%d%H%M%S")
     filename = os.path.basename(DEF_DBASE) + tstamp
     dest_file = os.path.join(AZOTEA_BAK_DIR, filename)
     shutil.copy2(DEF_DBASE, dest_file)
     logging.info("database backup to {0}".format(dest_file))
+
+
+def dbase_delete_selected_images(connection, batch):
+    row = {'batch': batch}
+    cursor = connection.cursor()
+    cursor.execute('''
+        DELETE FROM image_t
+        WHERE batch == :batch 
+        ''', row)
+
+def dbase_delete_selected_master_dark(connection, batch):
+    row = {'batch': batch}
+    cursor = connection.cursor()
+    cursor.execute('''
+        DELETE FROM master_dark_t
+        WHERE batch == :batch 
+        ''', row)
 
 # =====================
 # Command esntry points
@@ -54,10 +80,17 @@ def dbase_do_backup(ccomment):
 
 def database_clear(connection, options):
     cursor = connection.cursor()
-    cursor.execute("DELETE FROM image_t")
-    cursor.execute("DELETE FROM master_dark_t")
+    if options.all:
+        cursor.execute("DELETE FROM image_t")
+        cursor.execute("DELETE FROM master_dark_t")
+        logging.info("Cleared all data from database {0}".format(os.path.basename(DEF_DBASE)))
+    else:
+        batch = latest_batch(connection)
+        dbase_delete_selected_master_dark(connection, batch)
+        dbase_delete_selected_images(connection, batch)
+        logging.info("Cleared data from batch {1} in database {0}, ".format(os.path.basename(DEF_DBASE), batch))
     connection.commit()
-    logging.info("Cleared data from database {0}".format(os.path.basename(DEF_DBASE)))
+
 
 
 def database_purge(connection, options):
